changed: put Deviation struct in a separate header
ODR
