+++ examples/test_util/Deviation.hpp
@@ -1,0 +1,32 @@
+/*
+   Copyright 2016 Statoil ASA.
+
+   This file is part of the Open Porous Media project (OPM).
+
+   OPM is free software: you can redistribute it and/or modify
+   it under the terms of the GNU General Public License as published by
+   the Free Software Foundation, either version 3 of the License, or
+   (at your option) any later version.
+
+   OPM is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with OPM.  If not, see <http://www.gnu.org/licenses/>.
+   */
+
+#ifndef DEVIATION_HPP
+#define DEVIATION_HPP
+
+/*! \brief Deviation struct.
+    \details The member variables are default initialized to -1,
+             which is an invalid deviation value.
+ */
+struct Deviation {
+    double abs = -1; //!< Absolute deviation
+    double rel = -1; //!< Relative deviation
+};
+
+#endif
